+++ src/App.js
@@ -6,9 +6,10 @@ import Dashboard from './components/Dashboard/Dashboard';
 import AddNew from './components/AddNew/AddNew';
 import Home from './components/Home/Home';
 import Login from './components/Login/Login';
-import ShowRun from './components/ShowMore/ShowMore';
+import Signup from './components/Signup/Signup';
+import RunDetail from './components/RunDetail/RunDetail';
 import EditRun from './components/EditRun/EditRun';
-// import './App.css';
+// import './App.scss';
 
 class App extends Component {
   constructor() {
@@ -56,14 +57,17 @@ class App extends Component {
             <Route path="/login">
               <Login setUser={this.setUser} user={this.state.user} isLoggedIn={this.state.isLoggedIn} />
             </Route>
+            <Route path="/signup">
+              <Signup setUser={this.setUser} user={this.state.user} isLoggedIn={this.state.isLoggedIn} />
+            </Route>
             <PrivateRoute path="/users/:userId/runs/add" isLoggedIn={this.state.isLoggedIn}>
               <AddNew user={this.state.user} />
             </PrivateRoute>
             <PrivateRoute path="/users/:userId/runs/:runId/edit" isLoggedIn={this.state.isLoggedIn}>
-              <EditRun user={this.state.user} />
+              <EditRun user={this.state.user} run={this.state.selectedRun} setRun={this.setRun} />
             </PrivateRoute>
             <PrivateRoute path="/users/:userId/runs/:runId" isLoggedIn={this.state.isLoggedIn}>
-              <ShowRun user={this.state.user} run={this.state.selectedRun} setRun={this.setRun} />
+              <RunDetail user={this.state.user} run={this.state.selectedRun} setRun={this.setRun} />
             </PrivateRoute>
             <PrivateRoute path="/users/:userId" isLoggedIn={this.state.isLoggedIn}>
               <Dashboard user={this.state.user} setRun={this.setRun} selectedRun={this.state.selectedRun} />
