+++ src/components/DeleteRun/DeleteRun.scss
@@ -1,0 +1,20 @@
+.modal-outer {
+  background-color: rgba(0, 0, 0, 0.7);
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 10;
+}
+
+.modal-inner {
+  width: 50%;
+  max-width: 500px;
+  border-radius: 6px;
+  background-color: gainsboro;
+  padding: 2rem;
+}
