--- src/components/Login/LoginWrapper.css
@@ -1,45 +1,0 @@
-.login-modal {
-  background-color: rgba(0, 0, 0, 0.7);
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.login-wrapper {
-  width: 50%;
-  max-width: 500px;
-}
-
-.login-tabs {
-  text-align: right;
-  padding-right: 2.5rem;
-}
-.login-tabs button {
-  appearance: none;
-  cursor: pointer;
-  padding: 0.5rem 2rem;
-  background-color: darkslategray;
-  color: gainsboro;
-  border-radius: 6px 6px 0 0;
-  border: none;
-  font-size: 1.2rem;
-}
-.login-tabs button:hover,
-.login-tabs button:focus {
-  text-decoration: underline;
-}
-.login-tabs button.active {
-  background-color: gainsboro;
-  color: darkslategray;
-}
-
-.login-body {
-  border-radius: 6px;
-  background-color: gainsboro;
-  padding: 2rem;
-}
