+++ src/components/RunDetail/RunDetail.jsx
@@ -2,10 +2,10 @@ import React, { useState } from 'react';
 import { Link, Redirect } from 'react-router-dom';
 import Moment from 'react-moment';
 import moment from 'moment';
-import Delete from '../Delete/Delete';
-import './ShowMore.css';
+import DeleteRun from '../DeleteRun/DeleteRun';
+import './RunDetail.scss';
 
-const ShowMore = ({ user, run, setRun }) => {
+const RunDetail = ({ user, run, setRun }) => {
   const [showDelete, setShowDelete] = useState(false);
 
   const displayRun = run;
@@ -13,23 +13,11 @@ const ShowMore = ({ user, run, setRun }) => {
   let end = new moment(displayRun.end);
   displayRun.elapsedTime = moment.duration(end.diff(start));
 
-
-  const updateDisplay = (newRun) => {
-    displayRun.start = newRun.start;
-    displayRun.distance = newRun.distance;
-    displayRun.end = newRun.end;
-    start = new moment(newRun.start);
-    end = new moment(newRun.end);
-    displayRun.elapsedTime = moment.duration(end.diff(start));
-    displayRun.workoutType = newRun.workoutType;
-    displayRun.notes = newRun.notes;
-  }
-
   return (
     <>
       {!run._id ? <Redirect to={`/users/${user._id}`} /> : (
         <>
-          {showDelete && <Delete setRun={setRun} cancel={() => setShowDelete(false)} run={displayRun} />}
+          {showDelete && <DeleteRun setRun={setRun} closeModal={() => setShowDelete(false)} run={displayRun} />}
           <div className='show-more-wrapper'>
             <div className='close-wrap'>
               <button onClick={() => setRun({})}>Close</button>
@@ -41,7 +29,7 @@ const ShowMore = ({ user, run, setRun }) => {
             <p>Notes: {displayRun.notes}</p>
             <div>
               <button className='cancel' onClick={() => setShowDelete(true)}>Delete run</button>
-              <Link to={`users/${user._id}/runs/${run._id}/edit`}>Edit run</Link>
+              <Link to={`/users/${user._id}/runs/${run._id}/edit`}>Edit run</Link>
             </div>
           </div>
         </>
@@ -50,4 +38,4 @@ const ShowMore = ({ user, run, setRun }) => {
   );
 }
 
-export default ShowMore;
+export default RunDetail;
